+++ README_zh_CN.md
@@ -465,7 +465,7 @@ err = tx.SignEnvelope(account2.Address, key4.Index, key4Signer)
 You can submit a transaction to the network using the Access API client.
 
 ```go
-import "github.com/onflow/flow-go-sdk/client"
+import "github.com/onflow/flow-go-sdk/access"
 
 // 连接本地服务
 c, err := client.New("localhost:3569")
@@ -585,7 +585,7 @@ myID := ID.Int()
 你可以查询事件与' GetEventsForHeightRange '函数:
 
 ```go
-import "github.com/onflow/flow-go-sdk/client"
+import "github.com/onflow/flow-go-sdk/access"
 
 blocks, err := c.GetEventsForHeightRange(ctx, client.EventRangeQuery{
     Type:       "flow.AccountCreated",
