+++ access/http/client_test.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/onflow/flow-go/engine/access/rest/models"
 
-	"github.com/onflow/flow-go-sdk/client/convert"
+	"github.com/onflow/flow-go-sdk/access/convert"
 
 	"github.com/onflow/flow-go-sdk"
 	"github.com/onflow/flow-go-sdk/test"
