+++ examples/send_transactions/main.go
@@ -21,7 +21,7 @@ package main
 import (
 	"context"
 
-	"github.com/onflow/flow-go-sdk/client/http"
+	"github.com/onflow/flow-go-sdk/access/http"
 
 	"github.com/onflow/flow-go-sdk"
 	"github.com/onflow/flow-go-sdk/examples"
