+++ examples/verify_signature/user_signature_validate_all/main.go
@@ -23,7 +23,7 @@ import (
 	"encoding/hex"
 	"fmt"
 
-	"github.com/onflow/flow-go-sdk/client/http"
+	"github.com/onflow/flow-go-sdk/access/http"
 
 	"github.com/onflow/cadence"
 	"github.com/onflow/flow-go-sdk"
